% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assignLast}
\alias{assignLast}
\alias{asd}
\alias{zxc}
\title{Assign the most recent return (\code{.Last.value}) to a name}
\usage{
assignLast(name)

asd()

zxc()
}
\description{
Assign the most recent return (\code{.Last.value}) to a name
}
\section{Functions}{
\itemize{
\item \code{asd}: Assign .Last.value to asd

\item \code{zxc}: Assign .Last.value to asd
}}

