\name{swanr-backend-functions}
\alias{\%||\%}
\alias{swanr_cache_space}
\alias{swanr_cache_has}
\alias{swanr_cache_get}
\alias{swanr_cache_set}
\alias{swanr_hash_text}
\alias{swotr_auth_token}
\alias{swotr_authenticate}
\alias{swotr_has_credentials}
\alias{safe_filename}
\alias{safe_unzip}
\alias{read_spatial_file}
\alias{sf_to_geojson}
\alias{repair_layer_text}
\alias{clean_layer_attributes}
\alias{find_state_column}
\alias{load_brazil_state}
\alias{south_america_bbox}
\alias{reference_layer_specs}
\alias{load_reference_layer}
\alias{normalize_search_text}
\alias{catalog_rows}
\alias{layer_column}
\alias{build_layer_search_catalog}
\alias{search_layer_catalog}
\alias{make_bbox_sf}
\alias{iso_utc}
\alias{monthly_temporal_ranges}
\alias{cmr_get}
\alias{search_swot_data}
\alias{extract_download_link}
\alias{suffix_score}
\alias{select_best_versions}
\alias{entries_to_results}
\alias{product_configuration}
\alias{smart_filter_tiles}
\alias{build_granule_patterns}
\alias{validate_download_url}
\alias{download_earthdata_file}
\alias{clip_vector_zip}
\alias{clip_netcdf}
\alias{clip_other_vector}
\alias{ana_service_url}
\alias{ana_xml_table}
\alias{ANA.hidroinventario}
\alias{ana_daily_columns}
\alias{ana_daily_name}
\alias{ana_numeric}
\alias{serie.historica.ANA.proxy}
\alias{serie.historica.ANA}
\alias{parse_station_codes}
\alias{ana_subproduct_info}
\alias{station_code_vector}
\alias{select_ana_stations}
\alias{ana_inventory_from_station}
\alias{prepare_ana_results}
\alias{create_ana_download_bundle}
\alias{opera_product_configuration}
\alias{swanr_python}
\alias{run_python_script}
\alias{sword_node_columns}
\alias{sword_nodes_source}
\alias{utm_epsg_for_sf}
\alias{prepare_opera_sword_nodes}
\alias{swanr_output_dir}
\alias{opera_aoi_cache_key}
\alias{opera_cache_paths}
\alias{opera_file_table}
\alias{prepare_opera_results}
\alias{create_opera_download_bundle}
\alias{create_download_bundle}
\title{SWANr dashboard backend functions}
\description{
Functions used by the SWANr dashboard to query and prepare SWOT, ANA, and
OPERA data; manage spatial layers; and create download bundles.
}
\details{
These lower-level helpers are exported so they can be used independently with
\code{SWANr::}.  Functions that take \code{base_dir} expect the bundled app
directory, available from \code{system.file("app", package = "SWANr")}.
}
\keyword{utilities}
