% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mb}
\alias{mb}
\title{Convenience wrapper for \code{microbenchmark::microbenchmark}}
\usage{
mb(
  ...,
  list = NULL,
  times = 100L,
  unit,
  check = NULL,
  control = list(),
  setup = NULL
)
}
\arguments{
\item{...}{Expressions to benchmark.}

\item{list}{List of unevaluated expression to benchmark.}

\item{times}{Number of times to evaluate the expression.}

\item{unit}{Default unit used in \code{summary} and \code{print}.}

\item{check}{A function to check if the expressions are equal. By default \code{NULL} which omits the check.
In addition to a function, a string can be supplied.
The string \sQuote{equal} will compare all values using \code{\link{all.equal}}, \sQuote{equivalent} will compare all values using \code{\link{all.equal}} and check.attributes = FALSE, and \sQuote{identical} will compare all values using \code{\link{identical}}.}

\item{control}{List of control arguments. See Details.}

\item{setup}{An unevaluated expression to be run (untimed) before each benchmark expression.}
}
\description{
Convenience wrapper for \code{microbenchmark::microbenchmark}
}
