% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na.removeCols}
\alias{na.removeCols}
\title{Remove all columns that have any \code{NA}s}
\usage{
na.removeCols(x)
}
\arguments{
\item{x}{A matrix or data.frame}
}
\description{
Remove all columns that have any \code{NA}s
}
