% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{noneDefault}
\alias{noneDefault}
\title{Provide a default if a value is \code{NULL} or \code{NA}}
\usage{
noneDefault(value, default)
}
\description{
This is useful mostly when piping lists of data through map
functions. Often, one member of the pipeline will be NA or NULL where
an empty data.frame or xts is needed. This is a great for selectively
mapping that to the correct type.
}
