% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sdp}
\alias{sdp}
\title{Excel's STDEV.P}
\usage{
sdp(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an \R object but not a
    \code{\link{factor}} coercible to numeric by \code{as.double(x)}.}

\item{na.rm}{logical.  Should missing values be removed?}
}
\description{
This is the naive and biased estimator for standard deviation. It is often
(mis)used in excel. When collaborating with others it may be necessary to use.
}
