% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selfName}
\alias{selfName}
\title{Assign a vector's values to its names}
\usage{
selfName(x)
}
\arguments{
\item{x}{A vector}
}
\description{

}
\examples{
selfname(1:10)

c(4,6,8) \%>\%
selfName \%>\%
map(~ mtcars[mtcars$cyl == .x, ])

 @export
}
